+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - 0.1.0 (not yet released)
   - Added `parallelMap()` and `makeParallelMap()`.
   - Added `makeAsync()`.
+  - Made `makeFunc()` accept a context (`this`) object.
 
 - 0.0.0 (2012/02/19)
   - First release.