+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-- 0.0.1 (not yet released)
+- 0.0.1 (2012/02/25)
   - Renamed `series()` to `runSeq()`.
   - Renamed `makeSeries()` to `seq()`.
   - Renamed `parallel()` to `runPar()`.